Switch memory allocator, add new workflow
The new memory allocator should outperform most OS-provided allocators (which Dokkoo formerly used by default). This should increase binary size, and may lead to unexpected memory-related issues on certain platforms; however, I don't expect this to be much of a problem (if it is, it is treatable), and the performance increase is favourable.

The new workflow applies suggestions made by the Rust compiler; such suggestions are seen in the evaluation workflow—these will now be automatically applied.